Refactor: Renamed 'sushi' project to 'core'
The name of the project has been changed from 'sushi' to 'core'. As a
result, the environment settings in manage.py, asgi.py, settings.py,
urls.py, wsgi.py are updated to reflect this change. Also, the Gunicorn
WSGI http server call in Procfile has been updated. This change improves
the general understanding of the project's focus and helps with future
scalability.