refactor: update environment variables loading
Refactored the settings.py to use the python-dotenv library to load
environment variables, making it more standardized and robust. The DEBUG
was set to False when the DEVELEOPMENT environment variable is present
otherwise it's True. Also, enforced security settings in the production
environment such as XSS Filter, Content Type NoSniff, and others,
improving the security of the application. The requirements.txt was
updated with newer libraries' versions to keep the project up to date
with the latest improvements and fixes.